program/requirements.txt
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
pyinstaller==6.1.0
Pillow>=10.0.0
matplotlib>=3.7.0
cairosvg>=2.7.0

